netbox: add device_transition custom field (#582)
In this field for a device, the action/transition
that has just been performed can be recorded. E.g.
from_a_b if the device is currently in transition
from state a to state b. Or complete_a_b if the
transition from state a to state b has been completed.

Signed-off-by: Christian Berendt <berendt@osism.tech>