manager: make the number of netbox replicas configurable (#671)
This will start several Netbox Worker services. Per default 1.

Necessary for this is the use of Compose v2. This is provided
via the new CLI plugin for Docker.

With the old Compose CLI the key will be ignored. That's fine.

WARNING: Some services (netbox) use the 'deploy' key,
which will be ignored. Compose does not support 'deploy'
configuration - use `docker stack deploy` to deploy to
a swarm.

Signed-off-by: Christian Berendt <berendt@osism.tech>