Refactor wireguard role to allow multiple clients
In a production environment, we want to allow for multiple users to be
able to connect to the wireguard VPN simultaneously, so we add the
option to create a peer configuration for each of them.

Also adapt the role to running on recent Ubuntu versions, where the
module is available in the kernel by default.

Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>