manager: ignore warnings in "Copy private ssh keys"
Depending on the time at which the role is executed, certain keys are
not yet present in the correct places. However, this is fine and a warning
is therefore not necessary. So that the warning that a key has not yet
been found is not displayed, warnings are ignored here.

This will avoid the following message:

[WARNING]: Unable to find '/ansible/secrets/id_rsa.operator' in
expected paths (use -vvvvv to see paths)

Signed-off-by: Christian Berendt <berendt@osism.tech>