Default to non-interactive mode in wrapper scripts
Attaching a terminal during docker exec results in an additional
carriage return before newlines ([1]). While this is expected behaviour
by TTYs, this leads to unexpected results when piping output and
splitting on newlines.
Therefore we default to non-interactive mode if any of stdin, stdout and
stderr is not a terminal.

[1]
https://github.com/moby/moby/issues/37366

Signed-off-by: Jan Horstmann <horstmann@osism.tech>